Testing bean scopes
If we have stateless object best options to use is "singleton". This type of scope create only one object, and if we call from context.getBean() it will refer us to one object.
To make that clear if we create two object by getting bean and change something in first object, this  changes will also be applied to second object. If we have stateful object we have to use scope- "prototype"